Remove redundant fields from payments screen
Only keep the relevant parts which allow us to determine whether to
show or not show the payment details button
