+++ app/feature/feature-payments/src/main/kotlin/com/hedvig/android/feature/payments/data/PaymentConnection.kt
@@ -3,11 +3,7 @@ package com.hedvig.android.feature.payments.data
 import kotlinx.datetime.LocalDate
 
 internal sealed interface PaymentConnection {
-  data class Active(
-    val paymentMethod: PaymentMethod,
-    val chargingDay: Int?,
-    val account: PaymentAccount?,
-  ) : PaymentConnection
+  data object Active : PaymentConnection
 
   data object Pending : PaymentConnection
 
@@ -16,21 +12,4 @@ internal sealed interface PaymentConnection {
   ) : PaymentConnection
 
   data object Unknown : PaymentConnection
-}
-
-internal enum class PaymentMethod {
-  TRUSTLY,
-  SWISH,
-  NORDEA,
-  INVOICE,
-}
-
-internal sealed interface PaymentAccount {
-  data object Kivra : PaymentAccount
-
-  data class Email(val email: String) : PaymentAccount
-
-  data class PhoneNumber(val phoneNumber: String) : PaymentAccount
-
-  data class BankAccount(val account: String) : PaymentAccount
 }
